requirements.txt
runpod==1.7.7

# Qwen-Image-Edit-2511 needs a recent diffusers build
git+https://github.com/huggingface/diffusers.git@main

transformers>=4.44.2
accelerate>=0.33.0
safetensors>=0.4.5
huggingface_hub>=0.24.6

pillow>=10.4.0
numpy>=2.0.0
